+++ Trabalho Final/core/src/com/poo/jogo/Verde.java
@@ -6,7 +6,7 @@ public class Verde extends Especie {
 		super(pai);
 	}
 
-	Verde(int x, int y, float velocidade, float inteligencia, float tamanho, Tabuleiro tabuleiro) {
+	Verde(int x, int y, int velocidade, int inteligencia, int tamanho, Tabuleiro tabuleiro) {
 		super(x, y, velocidade, inteligencia, tamanho, tabuleiro, Cores.VERDE);
 	}
